Add files via upload
These are the xaml file of smart insurance processing - the details of the person from their personal identity certificates are collected with the help of OpenAI and get stored in the excel and the details automatically filled in the insurance form. The second projects send answer to the client query in different modes  like email, message, phone call.